the module now performs rain function
svn path=/trunk/iCub/; revision=2442

+++ src/saliencyBlobFinder/include/iCub/YARPBox.h
@@ -76,9 +76,9 @@ public:
 	int cGR;
 	int cBY;
 
-	int salienceBU;
-	int salienceTD;
-	int salienceTotal;
+	double salienceBU;
+	double salienceTD;
+	double salienceTotal;
 
 	double elev;
 	double az;
